+++ packaging/homebrew/Casks/ayame.rb
@@ -1,9 +1,9 @@
 cask "ayame" do
   arch arm: "aarch64", intel: "x86_64"
 
-  version "0.7.1"
-  sha256 arm:   "85fdba1e60809270d38547a3a1e493aa1e6eb81b65c88a6b8ee51e48bb34cbdf",
-         intel: "ec5f695f27c720245fbb488565a0c1b69c443643174d7c221907b37191c699b2"
+  version "0.7.2"
+  sha256 arm:   "1a0c98d96b4a940776246b079e60bf27fc743d0b8d541f2d7b9ce17c62fb6f2b",
+         intel: "ace2f9613287458e633cfb9a19ca90bbe675d938423695203a1b8669e02954d0"
 
   url "https://github.com/hjosugi/ayame-editor/releases/download/v#{version}/ayame-v#{version}-macos-#{arch}.zip"
   name "Ayame Editor"